+++ rules/downgrade-php71/src/Rector/FunctionLike/AbstractDowngradeRector.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Rector\DowngradePhp72\Rector\FunctionLike;
+namespace Rector\DowngradePhp71\Rector\FunctionLike;
 
 abstract class AbstractDowngradeRector extends AbstractMaybeAddDocBlockRector
 {
